Перенос основного расчета из main.cpp в glnsvpos.cpp. Перенос всех вспомогательных функций в func.cpp. В модели добавлен вывод разницы координат.

+++ libglnsvpos/include/libglnsvpos/structures.h
@@ -5,7 +5,7 @@
 
 struct Ephemeris_s {
     // Time in Gln
-    uint8_t   N4;
+    uint16_t   N4;
     uint16_t  NT;
     uint32_t  tb;
     // Coordinates
